move logic to service, fix cancel button
