+++ src/app/ubs/ubs-admin/components/shared/components/column-filters-pop-up/column-filters-pop-up.component.html
@@ -1,19 +1,16 @@
 <ng-container class="dropdown-menu">
   <div class="btn-group" *ngIf="showButtons">
-    <button type="button" class="btn filter-btn" [mat-dialog-close]="['clear', data.columnName]">
+    <button type="button" class="btn filter-btn" (click)="discardChanges()">
       {{ 'ubs-tables.clear-filters' | translate }}
     </button>
-    <button type="button" class="btn filter-btn bg-lime" [mat-dialog-close]="['apply', selectedFilters]">
+    <button type="button" class="btn filter-btn bg-lime" [mat-dialog-close]="['apply']">
       {{ 'ubs-tables.apply-filters' | translate }}
     </button>
   </div>
 
   <ng-container *ngIf="!data.isDateFilter; else date">
     <li *ngFor="let option of getOptionsForFiltering()" class="positioing">
-      <mat-checkbox
-        [checked]="isFilterChecked(data.columnName, option)"
-        (change)="changeColumnFilters($event?.checked, data.columnName, option)"
-      >
+      <mat-checkbox [checked]="isChecked(data.columnName, option)" (change)="onFilterChange($event?.checked, data.columnName, option)">
         {{ option | serverTranslate: data.currentLang }}
       </mat-checkbox>
     </li>
